requirements.txt
aiogram>=3.13,<3.14
requests>=2.32,<3
python-dotenv>=1.0.0